% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFDR.R
\name{BayesianFDR}
\alias{BayesianFDR}
\title{BayesianFDR}
\usage{
BayesianFDR(belief)
}
\arguments{
\item{belief}{Estilated probability of each gene to be intrinsic.}
}
\value{
A vector
\item{qvalue}{A vector of qvalues}
}
\description{
Get Bayesian FDR
}
\details{
Obtain Bayesian FDR after using GBclustering.
}
\author{
Lingsong Meng
}
