% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBclustering.R
\name{GBclustering}
\alias{GBclustering}
\title{GBclustering}
\usage{
GBclustering(
  Y,
  U,
  K,
  n.iter = 500,
  hyperparameters = c(1, 1, 1, 0.001, 0.001, 0.001, 0.001, 0.001, 0.001, 2, 0.005, 4,
    450),
  showIteration = T
)
}
\arguments{
\item{Y}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{U}{Adjusted R-squared or adjusted pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}

\item{K}{Number of clusters.}

\item{n.iter}{Number of iteration.}

\item{hyperparameters}{A vector including the following hyperparameters in order: c, a_p, b_p, a_sigma, b_sigma, a_U0, b_U0, a_U1, b_U1, a_mu0, b_mu0, a_mu1, b_mu1.}

\item{showIteration}{Output progress or not.}
}
\value{
a list consisting of
\item{L_PosterSamp}{Posterior samples of differential expression indicator L}
\item{Subtypes}{Clustering results}
\item{p}{Posterior mean of DE proportion p}
\item{tau_mu0}{Posterior mean of tau_mu0}
\item{tau_mu1}{Posterior mean of tau_mu1}
\item{tau_U0}{Posterior mean of tau_U0}
\item{tau_U1}{Posterior mean of tau_U1}
\item{pi}{Posterior mean of subtype proportions pi}
\item{mu}{Posterior mean of subtype mean mu}
\item{sigma_sq}{Posterior mean of subtype variance sigma_sq}
\item{tau_mu0_record}{Iteration record for tau_mu0}
\item{tau_mu1_record}{Iteration record for tau_mu1}
\item{tau_U0_record}{Iteration record for tau_U0}
\item{tau_U1_record}{Iteration record for tau_U1}
\item{mu_record}{Iteration record for mu}
\item{sigma_sq_record}{Iteration record for sigma_sq}
\item{BIC}{BIC value}
}
\description{
Guided Bayesian Clustering
}
\details{
Guided Bayesian clustering integrating clinical dataset with gene expression dataset.
}
\author{
Lingsong Meng
}
