% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getU.R
\name{getU}
\alias{getU}
\title{getU}
\usage{
getU(x, z, model)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{z}{One phenotypic variable from clinical dataset, a vector.}

\item{model}{The model fitted to obtain R2, please select model from 'linear', 'logit', 'exp', 'polr','cox'.}
}
\value{
A vector
\item{U}{adjusted R-squared or adjusted pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}
}
\description{
Get adjusted R-squared or adjusted pseudo R-squared
}
\details{
Obtain adjusted R-squared or adjusted pseudo R-squared between phenotypic variable and expression value of each gene,
which will be used in GBclustering function.
}
\author{
Lingsong Meng
}
